2023/CVE-2023-25355/pyproject.toml
[tool.poetry]
name = "cve-2023-25355-25356"
version = "0.1.2"
description = ""
authors = ["Guillem Lefait"]
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.scripts]
cve_2023_25355 = 'src.main:cve_2023_25355'
cve_2023_25356 = 'src.main:cve_2023_25356'
sipxcom_service_restart = 'src.main:restart_xmpp_service_with_superadmin'

[tool.poetry.dependencies]
python = "^3.12"
xmpppy = "^0.7.1"
click = "^8.1.7"
httpx = "^0.27.2"
beautifulsoup4 = "^4.12.3"

[tool.black]
line-length = 120
target-version = ['py312']
preview = true
skip-string-normalization = true

[tool.ruff]
line-length = 120
target-version = 'py312'
src = [
    'kelpri'
]

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pre-commit = "^3.7.0"
ruff = "^0.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
